requirements.txt
jinja2
gitpython
conda-build-all
pyyaml
